+++ source/MarketRoles.B2B.CimMessageAdapter.IntegrationTests/MessageReceiverTests.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Xml;
-using System.Xml.Schema;
 using B2B.CimMessageAdapter;
 using Xunit;
 
@@ -71,7 +67,7 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
 
             var activityRecord = _marketActivityRecordForwarderStub.CommittedItems.FirstOrDefault();
             Assert.NotNull(activityRecord);
-            Assert.Equal("12345699", activityRecord.MRid);
+            Assert.Equal("12345699", activityRecord.MrId);
             Assert.Equal("579999993331812345", activityRecord.MarketEvaluationPointmRID);
             Assert.Equal("5799999933318", activityRecord.EnergySupplierMarketParticipantmRID);
             Assert.Equal("5799999933340", activityRecord.BalanceResponsiblePartyMarketParticipantmRID);
@@ -147,233 +143,6 @@ namespace MarketRoles.B2B.CimMessageAdapter.IntegrationTests
             fileReader.CopyTo(messageStream);
             messageStream.Position = 0;
             return messageStream;
-        }
-    }
-
-    public class MessageReceiver
-    {
-        private readonly List<ValidationError> _errors = new();
-        private readonly IMessageIds _messageIds;
-        private readonly IMarketActivityRecordForwarder _marketActivityRecordForwarder;
-        private readonly ITransactionIds _transactionIds;
-        private bool _hasInvalidHeaderValues;
-
-        public MessageReceiver(IMessageIds messageIds, IMarketActivityRecordForwarder marketActivityRecordForwarder, ITransactionIds transactionIds)
-        {
-            _messageIds = messageIds ?? throw new ArgumentNullException(nameof(messageIds));
-            _marketActivityRecordForwarder = marketActivityRecordForwarder ?? throw new ArgumentNullException(nameof(marketActivityRecordForwarder));
-            _transactionIds = transactionIds;
-        }
-
-        public async Task<Result> ReceiveAsync(Stream message, string businessProcessType, string version)
-        {
-            if (message == null) throw new ArgumentNullException(nameof(message));
-
-            var xmlSchema = await GetSchemaAsync(businessProcessType, version).ConfigureAwait(true);
-            if (xmlSchema is null)
-            {
-                return Result.Failure(new ValidationError(
-                    $"Schema version {version} for business process type {businessProcessType} does not exist."));
-            }
-
-            _hasInvalidHeaderValues = false;
-            using (var reader = XmlReader.Create(message, CreateXmlReaderSettings(xmlSchema)))
-            {
-                try
-                {
-                    await HandleMessageHeaderValues(reader).ConfigureAwait(false);
-                    await HandleMarketActivityRecordsAsync(reader).ConfigureAwait(false);
-                }
-                catch (XmlException exception)
-                {
-                    return Result.Failure(new ValidationError(exception.Message));
-                }
-            }
-
-            if (_hasInvalidHeaderValues == false)
-            {
-                await _marketActivityRecordForwarder.CommitAsync().ConfigureAwait(false);
-            }
-            return _errors.Count == 0 ? Result.Succeeded() : Result.Failure(_errors.ToArray());
-        }
-
-        private async Task HandleMarketActivityRecordsAsync(XmlReader reader)
-        {
-            while (await reader.ReadAsync())
-            {
-                if (reader.NodeType != XmlNodeType.Element || !reader.LocalName.Equals("MktActivityRecord")) continue;
-                await foreach (var marketActivityRecord in ExtractFromAsync(reader))
-                {
-                    if (await CheckTransactionIdAsync(marketActivityRecord.MRid).ConfigureAwait(false) == false)
-                    {
-                        _errors.Add(new ValidationError(
-                            $"Transaction id '{marketActivityRecord.MRid}' is not unique and will not be processed."));
-                    }
-                    else
-                    {
-                        await StoreActivityRecordAsync(marketActivityRecord).ConfigureAwait(false);
-                    }
-                }
-            }
-        }
-
-        private async IAsyncEnumerable<MarketActivityRecord> ExtractFromAsync(XmlReader reader)
-        {
-            var mRID = string.Empty;
-            var marketEvaluationPointmRID = string.Empty;
-            var energySupplierMarketParticipantmRID = string.Empty;
-            var balanceResponsiblePartyMarketParticipantmRID = string.Empty;
-            var customerMarketParticipantmRID = string.Empty;
-            var customerMarketParticipantname = string.Empty;
-            var startDateAndOrTimedateTime = string.Empty;
-
-            while (await reader.ReadAsync())
-            {
-                if (reader.NodeType == XmlNodeType.EndElement &&
-                    reader.LocalName.Equals("MktActivityRecord"))
-                {
-                    if (reader.SchemaInfo.Validity == XmlSchemaValidity.Invalid)
-                    {
-                        continue;
-                    }
-
-                    yield return new MarketActivityRecord()
-                    {
-                        MRid = mRID,
-                        CustomerMarketParticipantName = customerMarketParticipantname,
-                        CustomerMarketParticipantmRID = customerMarketParticipantmRID,
-                        MarketEvaluationPointmRID = marketEvaluationPointmRID,
-                        EnergySupplierMarketParticipantmRID = energySupplierMarketParticipantmRID,
-                        StartDateAndOrTimeDateTime = startDateAndOrTimedateTime,
-                        BalanceResponsiblePartyMarketParticipantmRID =
-                            balanceResponsiblePartyMarketParticipantmRID,
-                    };
-                }
-
-                if (reader.NodeType == XmlNodeType.Element)
-                {
-                    TryExtractValueFrom("mRID", reader, ref mRID);
-                    TryExtractValueFrom("marketEvaluationPoint.mRID", reader, ref marketEvaluationPointmRID);
-                    TryExtractValueFrom("marketEvaluationPoint.energySupplier_MarketParticipant.mRID", reader, ref energySupplierMarketParticipantmRID);
-                    TryExtractValueFrom("marketEvaluationPoint.balanceResponsibleParty_MarketParticipant.mRID", reader, ref balanceResponsiblePartyMarketParticipantmRID);
-                    TryExtractValueFrom("marketEvaluationPoint.customer_MarketParticipant.mRID", reader, ref customerMarketParticipantmRID);
-                    TryExtractValueFrom("marketEvaluationPoint.customer_MarketParticipant.name", reader, ref customerMarketParticipantname);
-                    TryExtractValueFrom("start_DateAndOrTime.dateTime", reader, ref startDateAndOrTimedateTime);
-                }
-            }
-        }
-
-        private void TryExtractValueFrom(string elementName, XmlReader reader, ref string variable)
-        {
-            if (reader.LocalName.Equals(elementName))
-            {
-                variable = reader.ReadElementString();
-            }
-        }
-
-        private async Task HandleMessageHeaderValues(XmlReader reader)
-        {
-            while (await reader.ReadAsync())
-            {
-                if (reader.NodeType == XmlNodeType.Element &&
-                    reader.LocalName.Equals("RequestChangeOfSupplier_MarketDocument"))
-                {
-                    while (await reader.ReadAsync())
-                    {
-                        if (reader.NodeType == XmlNodeType.Element && reader.LocalName.Equals("mRID"))
-                        {
-                            var messageId = reader.ReadElementString();
-                            var messageIdIsUnique = await CheckMessageIdAsync(messageId);
-                            if (messageIdIsUnique == false)
-                            {
-                                _errors.Add(new DuplicateId($"Message id '{messageId}' is not unique"));
-                                _hasInvalidHeaderValues = true;
-                            }
-
-                            break;
-                        }
-                    }
-
-                    break;
-                }
-            }
-        }
-
-        private Task<bool> CheckTransactionIdAsync(string transactionId)
-        {
-            if (transactionId == null) throw new ArgumentNullException(nameof(transactionId));
-            return _transactionIds.TryStoreAsync(transactionId);
-        }
-
-        private Task StoreActivityRecordAsync(MarketActivityRecord marketActivityRecord)
-        {
-            return _marketActivityRecordForwarder.AddAsync(marketActivityRecord);
-        }
-
-        private Task<bool> CheckMessageIdAsync(string messageId)
-        {
-            if (messageId == null) throw new ArgumentNullException(nameof(messageId));
-            return _messageIds.TryStoreAsync(messageId);
-        }
-
-        private XmlReaderSettings CreateXmlReaderSettings(XmlSchema xmlSchema)
-        {
-            var settings = new XmlReaderSettings
-            {
-                Async = true,
-                ValidationType = ValidationType.Schema,
-                ValidationFlags = XmlSchemaValidationFlags.ProcessInlineSchema |
-                                  XmlSchemaValidationFlags.ReportValidationWarnings,
-            };
-
-            settings.Schemas.Add(xmlSchema);
-            settings.ValidationEventHandler += OnValidationError;
-            return settings;
-        }
-
-        private static Task<XmlSchema?> GetSchemaAsync(string businessProcessType, string version)
-        {
-            var schemas = new Dictionary<KeyValuePair<string, string>, string>()
-            {
-                {
-                    new KeyValuePair<string, string>("requestchangeofsupplier", "1.0"),
-                    "urn-ediel-org-structure-requestchangeofsupplier-0-1.xsd"
-                }
-            };
-
-            if (schemas.TryGetValue(new KeyValuePair<string, string>(businessProcessType, version),
-                    out var schemaName) == false)
-            {
-                return Task.FromResult(default(XmlSchema));
-            }
-
-            return LoadSchemaWithDependentSchemasAsync(schemaName);
-        }
-
-        private static async Task<XmlSchema> LoadSchemaWithDependentSchemasAsync(string location)
-        {
-            using var reader = new XmlTextReader(location);
-            var xmlSchema = XmlSchema.Read(reader, null);
-
-            foreach (XmlSchemaExternal external in xmlSchema.Includes)
-            {
-                if (external.SchemaLocation == null)
-                {
-                    continue;
-                }
-
-                external.Schema =
-                    await LoadSchemaWithDependentSchemasAsync(external.SchemaLocation).ConfigureAwait(false);
-            }
-
-            return xmlSchema;
-        }
-
-        private void OnValidationError(object? sender, ValidationEventArgs arguments)
-        {
-            var message =
-                $"XML schema validation error at line {arguments.Exception.LineNumber}, position {arguments.Exception.LinePosition}: {arguments.Message}.";
-            _errors.Add(new ValidationError(message));
         }
     }
 }
